+++ app/config.py
@@ -1,0 +1,4 @@
+DB_PATH = "stations20260220.db"
+STATIONS_URL = "https://services.drova.io/server-manager/servers/public/web"
+PRODUCTS_URL = "https://services.drova.io/product-manager/product/listfull2"
+CACHE_TTL_SECONDS = 600
